Isoler le worker QGIS
